requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
Flask==2.3.3
fsspec==2025.7.0
google==3.0.0
google-auth==2.40.3
google-genai==1.27.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==4.9.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx
nltk==3.9.1
numpy
openai==1.97.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.31
requests==2.31.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy
selenium==4.15.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webdriver-manager==4.0.1
websockets==15.0.1
Werkzeug==2.3.7
wsproto==1.2.0
